added models Athlete and coach along with their respective repo's.

+++ bmx-app-backend/bmx-app-backend/src/main/java/com/example/bmxappbackend/repository/TrackTimeRecordRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TrackTimeRecordRepository extends JpaRepository<TrackTimeRecord, Long> {
+public interface TrackTimeRecordRepository extends JpaRepository<TrackTimeRecord, Integer> {
 
     List<TrackTimeRecord> findAll();
